+++ src/openxr/OpenXRApi.h
@@ -92,14 +92,20 @@ public:
 		// Analog
 		ACTION_FRONT_TRIGGER, // front trigger (Axis 2)
 		ACTION_SIDE_TRIGGER, // side trigger/grip (Axis 4)
-		ACTION_JOYSTICK, // primary joystick (Axis 0/1)
+		ACTION_PRIMARY, // primary joystick/trackpad/thumbstick (Axis 0/1)
+		ACTION_SECONDARY, // secondary joystick/trackpad/thumbstick (Axis 6/7)
 
 		// Buttons
 		ACTION_AX_BUTTON, // A/X button (button 7)
-		ACTION_BYM_BUTTON, // B/Y/Menu button (button 1)
+		ACTION_BY_BUTTON, // B/Y button (button 1)
+		ACTION_MENU_BUTTON, // Menu button (button 3)
+		ACTION_SELECT_BUTTON, // Menu button (button 4)
 		ACTION_FRONT_BUTTON, // front trigger as button (button 15)
 		ACTION_SIDE_BUTTON, // side trigger/grip as button (button 2)
-		ACTION_JOYSTICK_BUTTON, // Click on joystick (button 14)
+		ACTION_PRIMARY_BUTTON, // Click on primary joystick/trackpad/thumbstick (button 14)
+		ACTION_SECONDARY_BUTTON, // Click on secondary joystick/trackpad/thumbstick (button 13)
+		ACTION_PRIMARY_TOUCH, // Finger is on primary joystick/trackpad/thumbstick (button 12)
+		ACTION_SECONDARY_TOUCH, // Finger is on secondary joystick/trackpad/thumbstick (button 12)
 
 		// Output
 		ACTION_HAPTIC, // Haptic output
@@ -118,12 +124,18 @@ public:
 		{ "grip_pose", XR_ACTION_TYPE_POSE_INPUT, NULL },
 		{ "front_trigger", XR_ACTION_TYPE_FLOAT_INPUT, NULL },
 		{ "side_trigger", XR_ACTION_TYPE_FLOAT_INPUT, NULL },
-		{ "joystick", XR_ACTION_TYPE_VECTOR2F_INPUT, NULL },
+		{ "primary", XR_ACTION_TYPE_VECTOR2F_INPUT, NULL },
+		{ "secondary", XR_ACTION_TYPE_VECTOR2F_INPUT, NULL },
 		{ "ax_buttons", XR_ACTION_TYPE_BOOLEAN_INPUT, NULL },
-		{ "bym_button", XR_ACTION_TYPE_BOOLEAN_INPUT, NULL },
+		{ "by_button", XR_ACTION_TYPE_BOOLEAN_INPUT, NULL },
+		{ "menu_button", XR_ACTION_TYPE_BOOLEAN_INPUT, NULL },
+		{ "select_button", XR_ACTION_TYPE_BOOLEAN_INPUT, NULL },
 		{ "front_button", XR_ACTION_TYPE_BOOLEAN_INPUT, NULL },
 		{ "side_button", XR_ACTION_TYPE_BOOLEAN_INPUT, NULL },
-		{ "joystick_button", XR_ACTION_TYPE_BOOLEAN_INPUT, NULL },
+		{ "primary_button", XR_ACTION_TYPE_BOOLEAN_INPUT, NULL },
+		{ "secondary_button", XR_ACTION_TYPE_BOOLEAN_INPUT, NULL },
+		{ "primary_touch", XR_ACTION_TYPE_BOOLEAN_INPUT, NULL },
+		{ "secondary_touch", XR_ACTION_TYPE_BOOLEAN_INPUT, NULL },
 		{ "haptic", XR_ACTION_TYPE_VIBRATION_OUTPUT, NULL },
 	};